Extract Practice component
